+++ modules/app-core/src/commonMain/kotlin/org/archivekeep/app/core/domain/storages/StorageDriver.kt
@@ -4,7 +4,6 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.SharedFlow
 import org.archivekeep.app.core.domain.repositories.RepoAuthRequest
 import org.archivekeep.app.core.utils.ProtectedLoadableResource
-import org.archivekeep.app.core.utils.generics.OptionalLoadable
 import org.archivekeep.app.core.utils.identifiers.RepositoryURI
 import org.archivekeep.app.core.utils.identifiers.StorageURI
 import org.archivekeep.files.repo.Repo
@@ -18,6 +17,6 @@ interface StorageDriver {
 
 data class StorageConnection(
     val storageURI: StorageURI,
-    val information: Flow<OptionalLoadable<StorageInformation>>,
+    val information: StorageInformation,
     val connectionStatus: SharedFlow<Loadable<Storage.ConnectionStatus>>,
 )