+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/desktop/ui/views/View.kt
@@ -4,13 +4,13 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import kotlinx.coroutines.CoroutineScope
 
-interface View<S> {
+interface View<VM> {
     @Composable
-    fun producePersistentState(scope: CoroutineScope): S
+    fun produceViewModel(scope: CoroutineScope): VM
 
     @Composable
     fun render(
         modifier: Modifier,
-        state: S,
+        vm: VM,
     )
 }