+++ modules/app-ui/src/commonMain/kotlin/org/archivekeep/app/desktop/ui/views/settings/SettingsView.kt
@@ -17,12 +17,12 @@ import org.archivekeep.app.desktop.ui.views.View
 
 class SettingsView : View<String> {
     @Composable
-    override fun producePersistentState(scope: CoroutineScope): String = "TODO"
+    override fun produceViewModel(scope: CoroutineScope): String = "TODO"
 
     @Composable
     override fun render(
         modifier: Modifier,
-        state: String,
+        vm: String,
     ) {
         Surface(
             modifier = modifier,